[NET-956] Refactor docapi module for none settings routes with builder support
