+++ app/src/app/docapi/docapi-builder/docapi-builder-type-step/docapi-builder-type-step-object/docapi-builder-type-step-object.component.ts
@@ -1,6 +1,6 @@
 /*
 * DATAGERRY - OpenSource Enterprise CMDB
-* Copyright (C) 2019 NETHINKS GmbH
+* Copyright (C) 2019 - 2021 NETHINKS GmbH
 *
 * This program is free software: you can redistribute it and/or modify
 * it under the terms of the GNU Affero General Public License as
@@ -13,38 +13,35 @@
 * GNU Affero General Public License for more details.
 
 * You should have received a copy of the GNU Affero General Public License
-* along with this program.  If not, see <https://www.gnu.org/licenses/>.
+* along with this program. If not, see <https://www.gnu.org/licenses/>.
 */
 
-import { Component, OnInit, Input } from '@angular/core';
-import { AbstractControl, FormControl, FormGroup, Validators } from '@angular/forms';
-import { CmdbType } from '../../../../../framework/models/cmdb-type';
-import { TypeService } from '../../../../../framework/services/type.service';
+import { Component, OnInit } from '@angular/core';
+import { FormControl, FormGroup, Validators } from '@angular/forms';
 import {
-  DocapiSettingsBuilderTypeStepBaseComponent
-} from '../docapi-settings-builder-type-step-base/docapi-settings-builder-type-step-base.component';
+  DocapiBuilderTypeStepBaseComponent
+} from '../docapi-builder-type-step-base/docapi-builder-type-step-base.component';
+import { TypeService } from '../../../../framework/services/type.service';
+import { CmdbType } from '../../../../framework/models/cmdb-type';
 
 @Component({
   selector: 'cmdb-docapi-settings-builder-type-step-object',
-  templateUrl: './docapi-settings-builder-type-step-object.component.html',
-  styleUrls: ['./docapi-settings-builder-type-step-object.component.scss']
+  templateUrl: './docapi-builder-type-step-object.component.html',
+  styleUrls: ['./docapi-builder-type-step-object.component.scss']
 })
-export class DocapiSettingsBuilderTypeStepObjectComponent extends DocapiSettingsBuilderTypeStepBaseComponent implements OnInit {
+export class DocapiBuilderTypeStepObjectComponent extends DocapiBuilderTypeStepBaseComponent implements OnInit {
 
 
-  public objectTypeList: CmdbType[] = [];
-
+  public objectTypeList: Array<CmdbType> = [];
 
   constructor(private typeService: TypeService) {
     super();
-    // setup form
     this.typeParamForm = new FormGroup({
       type: new FormControl('', Validators.required)
     });
   }
 
   public ngOnInit(): void {
-    // load object type list
     this.typeParamForm.valueChanges.subscribe(() => {
       this.formValid = this.typeParamForm.valid;
       this.formValidationEmitter.emit(this.formValid);
