+++ app/src/app/docapi/docapi-list/docapi-list.component.html
@@ -16,7 +16,7 @@
 <ng-template #actionsTemplate let-item="item">
   <div class="d-flex justify-content-center">
     <a class="ml-1 text-dark mr-1" *permissionLink="['base.docapi.template.edit']"
-       [routerLink]="['/settings/docapi/edit/', item?.public_id]">
+       [routerLink]="['/docapi/edit/', item?.public_id]">
       <fa-icon [icon]="['far', 'edit']"></fa-icon>
     </a>
     <a class="ml-1 text-dark mr-1" *permissionLink="['base.docapi.template.delete']">
@@ -29,7 +29,7 @@
   <button type="button"
           *permissionLink="['base.docapi.template.add']"
           class="btn btn-success btn-sm mr-1"
-          [routerLink]="['/settings/docapi/add']">
+          [routerLink]="['/docapi/add']">
     <i class="fas fa-plus"></i> Add
   </button>
 </ng-template>
